Hash out basic structure (work in progress)
The basic logic is there, but the integration test reveals the current
approach won't quite work. Intead of pushing on a Send handler before
and after the SDK's core Send logic, we need to REPLACE the core Send
logic with a wrapper handler. Within the wrapper we will:

- duplicate the HTTP request by making a shallow copy with a new context
  that contains the timeout;
- send the request by invoking the wrapped core Send logic; and
- swap back the old HTTP request.

The reason the current approach doesn't work, and the approach described
above will, is that HTTP request context is "sticky" once set. The AWS
SDK shallow copies the HTTP request structure on every retry, but the
shallow copy includes a copied reference to the previous context. The
impact of the copying behavior is that all subsequent timeouts for all
retries are limited by whatever deadline was put into the context
earlier. A child context can never have an earlier deadline than the
parent.

To work around this, we need to structure the code so that the SDK never
sees the version of the HTTP request