Change approach and finish fully working and tested implementation
As mentioned in the previous commit, in order to avoid making the
deadline "sticky" on the http.Request within the AWS SDK request, I had
to wrap the AWS SDK for Go's core Send handler and shallow-copy the HTTP
request.

As of this commit, we have a full working implementation with 100% test
coverage. All that's left now is to setup the Go module, finish the
documentation, etc.