Fix bug where context canceled during body read
Before this bug fix, flextime would work well for empty or "short"
bodies which could be fully buffered on the client-side before the
wrapped Core send handler returned. If, however, the body did not
get fully read by this point, and additional reads on the HTTP
connection were required to get the rest of the body, the connection
would get canceled by the `defer cancel()`, and this would blow up
the AWS SDK request that was relying on that connection.

After this bug fix, we simply do not cancel the child context we
create to manage the timeout. Since the context is a child of the
previous context, it will still be canceled if the parent context is
canceled, so higher-up request scoped control still works.